created a new file process_scrapwebio.py in this file created write_csv(header,data,filename) function to write csv files from lists.
header:  is a list contraining strings that are the header of the
csv file

data: is a list of lists where each list consists of the data for one
row in the csv file.

filename: is the full path to the csv file which will be created.

An example of a call (excuted when running  file process_scrapwebio.py):

header = ['Name', 'Course1 Score', 'Course2 Score']
data = [['Ali', 92, 80], ['Jonny', 75, 56], ['Mikey', 85, 98]]
filemame="./studentscore.csv"
write_csv(header,data,filename)